(move) handlers definitions to their own module
